Start refactor of queries and mutations into DB and API objects, removes the need to import await dbClient as well as should be a better user experience keeping track of queries and mutations on the server and client side
