+++ src/components/dashboard/WorkflowInstances.tsx
@@ -1,17 +1,16 @@
-import { workflowInstancesQueryOptions } from "@/data/workflowInstances";
 import { useSuspenseQuery } from "@tanstack/react-query";
 import { Link } from "@tanstack/react-router";
-import type { WorkflowInstance } from "@/data/workflowInstances";
+import { API } from "@/data/API";
 
 export function WorkflowInstances() {
 	const workflowInstancesQuery = useSuspenseQuery(
-		workflowInstancesQueryOptions(),
+		API.workflowInstance.queries.getWorkflowInstancesQueryOptions(),
 	);
 
 	return (
 		<section className="mb-8">
 			<h2 className="text-xl font-semibold mb-4">Recent Workflow Instances</h2>
-			{workflowInstancesQuery.data.map((instance: WorkflowInstance) => (
+			{workflowInstancesQuery.data.map((instance) => (
 				<div key={instance.id} className="mb-4 p-4 border rounded">
 					<div>
 						ID:{" "}
