+++ src/data/DB/index.ts
@@ -1,0 +1,5 @@
+import { workflowInstance } from "./workflowInstance";
+
+export const DB = {
+	workflowInstance,
+};
