+++ src/routes/workflowInstances/index.tsx
@@ -1,18 +1,20 @@
-import { workflowInstancesQueryOptions } from "@/data/workflowInstances";
 import { useSuspenseQuery } from "@tanstack/react-query";
 import { createFileRoute } from "@tanstack/react-router";
+import { API } from "@/data/API";
 
 export const Route = createFileRoute("/workflowInstances/")({
 	component: RouteComponent,
 	loader: async ({ context }) => {
-		return context.queryClient.ensureQueryData(workflowInstancesQueryOptions());
+		return context.queryClient.ensureQueryData(
+			API.workflowInstance.queries.getWorkflowInstancesQueryOptions(),
+		);
 	},
 	pendingComponent: () => <div>Loading...</div>,
 });
 
 function RouteComponent() {
 	const workflowInstancesQuery = useSuspenseQuery(
-		workflowInstancesQueryOptions(),
+		API.workflowInstance.queries.getWorkflowInstancesQueryOptions(),
 	);
 
 	return (
